ivi-id-agent-modules: fix weston header link
The WESTON_INCLUDE_DIRS contains a link to ${install_dir}/include/weston.
So, include the weston headers should not add the weston in the first path.

If user only build weston, the wayland and pixman are
installed via apt, they will got the missing weston header and
can not find the libweston-desktop.

Signed-off-by: Tran Ba Khang(MS/EMC31-XC) <Khang.TranBa@vn.bosch.com>